notifications/email-api/requirements.txt
aio-pika==9.4.3
fastapi==0.111.0
gunicorn==23.0.0
pydantic-settings==2.3.4
redis==5.0.4
psycopg2-binary==2.9.9